Attempting the reml regression used by afni programmed in Stan
Afni uses a reml regression where autocorrelation accounted for by a covariance matrix. I chose to use a high-pass filter instead of the polort regression coefficients because finding posteriors for those coefficients seem unhelpful. But perhaps a downward slope in the data like the ssVEP could be found that way. Nonetheless, model003 with a normal covariance matrix seems to do just fine. The correlation of two neighboring TRs is equal to delta * rho  ^ 1. The posterior of delta (mean = 0.5, SD =0.06) and rho (mean = 0.68, SD = 0.06) would form a solid correlation at .34 most of the time, so that sampler converges fine because that is a fairly symmetrical bowl for the sampler. But it is fairly slow, took 7 hours for 4 chains on a laptop of 1000 warmup and posterior samples. It is probably because of the large covariance matrix that I am making, or maybe multinormals are just slow with so many dimensions. But I saved the warmup samples and it didn't look like it needed that long to settling into posterior of each parameter. So I'm looking into running more chains at once but for shorter and then using the nest-R^2 measure of convergence by combining the little chains into super-chains haha, but it is justified based on Margossian et al., 2024; https://doi.org/10.1214/24-BA1453. They show with simulations that just one sample from 512 chains had basically the same error as 1000 samples from 4 chains. So if the sampler looked like it was working at 200 samples, then 16 chains of 200 will probably be justified. We can see if that is adequate for the size of these covariance matrices. May need to use map_rect to try to get it running faster.