Gaussian process model and prep
Added a script to get the design matrix per stimulus, but organized by block, cue, then shock. Updated prep script for stan so that the design matrix and list is formed per person. Model 26 is the first Gaussian process model. It creates betas for cues and shocks shared by all participants. So it is kind of like a moving average for the hemodynamic response. Perhaps it could be extended to allow a guassian process per participant, but it is so complex that I am starting small to see how it does versus model 18 that is more of a standard cue by phase. Model 26 still allows for time autocorrelated error per person. Model 26 also does not allow for variation beyond 1 ROI, just too complicated to think about right now. This will go on the cluster to see how it does on the 24 participant anterior insula vs model 18.